refactor(companion): group shared code by domain
