+++ companion/apps/android/src/main/java/com/rsvpnano/android/net/AndroidNanoNetworkController.kt
@@ -11,12 +11,12 @@ import android.net.nsd.NsdServiceInfo
 import android.net.wifi.WifiInfo
 import android.net.wifi.WifiNetworkSpecifier
 import android.os.PatternMatcher
-import com.rsvpnano.app.NanoWifiConnector
-import com.rsvpnano.app.NanoEndpoint
-import com.rsvpnano.app.NanoWifiEvent
-import com.rsvpnano.app.NanoWifiIdentity
-import com.rsvpnano.app.NanoWifiRequestResult
-import com.rsvpnano.app.NanoWifiSnapshot
+import com.rsvpnano.connection.NanoWifiConnector
+import com.rsvpnano.connection.NanoEndpoint
+import com.rsvpnano.connection.NanoWifiEvent
+import com.rsvpnano.connection.NanoWifiIdentity
+import com.rsvpnano.connection.NanoWifiRequestResult
+import com.rsvpnano.connection.NanoWifiSnapshot
 import com.rsvpnano.models.RememberedNano
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
