+++ companion/shared/build.gradle.kts
@@ -59,6 +59,7 @@ kotlin {
 
 		commonTest.dependencies {
 			implementation(kotlin("test"))
+			implementation(libs.kotlinx.coroutines.test)
 		}
 
 		androidMain.dependencies {
