+++ companion/shared/src/commonMain/kotlin/com/rsvpnano/library/ImportPreparation.kt
@@ -1,6 +1,8 @@
-package com.rsvpnano.converters
+package com.rsvpnano.library
 
 import com.rsvpnano.models.PendingUpload
+import com.rsvpnano.converters.RsvpBookFile
+import com.rsvpnano.converters.RsvpConverter
 
 object ImportPreparation {
     fun titleForText(preferredTitle: String, text: String, fallback: String): String {
