+++ companion/shared/src/commonMain/kotlin/com/rsvpnano/library/PendingDraftService.kt
@@ -1,4 +1,4 @@
-package com.rsvpnano.app
+package com.rsvpnano.library
 
 import com.rsvpnano.api.ArticleFetchClient
 import com.rsvpnano.api.NanoApi
