+++ companion/shared/src/commonMain/kotlin/com/rsvpnano/presentation/CompanionPresenter.kt
@@ -1,22 +1,23 @@
-package com.rsvpnano.ui
+package com.rsvpnano.presentation
 
 import com.rsvpnano.app.CompanionNotice
 import com.rsvpnano.app.CompanionCatalogFile
-import com.rsvpnano.app.FirmwareUpdates
+import com.rsvpnano.updates.FirmwareUpdates
 import com.rsvpnano.app.NanoCompanionController
-import com.rsvpnano.app.NanoConnectionState
-import com.rsvpnano.app.NanoConnectionTransport
-import com.rsvpnano.app.NanoEndpoint
+import com.rsvpnano.connection.NanoConnectionState
+import com.rsvpnano.connection.NanoConnectionTransport
+import com.rsvpnano.connection.NanoEndpoint
 import com.rsvpnano.app.NanoSettingsResource
-import com.rsvpnano.app.NanoWifiConnector
-import com.rsvpnano.app.NanoWifiEvent
-import com.rsvpnano.app.NanoWifiIdentity
-import com.rsvpnano.app.NanoWifiRequestResult
-import com.rsvpnano.app.NanoWifiSnapshot
+import com.rsvpnano.connection.NanoWifiConnector
+import com.rsvpnano.connection.NanoWifiEvent
+import com.rsvpnano.connection.NanoWifiIdentity
+import com.rsvpnano.connection.NanoWifiRequestResult
+import com.rsvpnano.connection.NanoWifiSnapshot
 import com.rsvpnano.app.SharedAppUtils
-import com.rsvpnano.app.catalogContentUrl
-import com.rsvpnano.app.releaseSource
-import com.rsvpnano.converters.ImportPreparation
+import com.rsvpnano.updates.catalogContentUrl
+import com.rsvpnano.updates.releaseSource
+import com.rsvpnano.library.ImportPreparation
+import com.rsvpnano.ui.library.isArticle
 import com.rsvpnano.converters.RsvpConverter
 import com.rsvpnano.models.NanoBook
 import com.rsvpnano.models.NanoLanguageFont
@@ -27,7 +28,7 @@ import com.rsvpnano.models.NanoSettingsSchema
 import com.rsvpnano.models.NanoWifiSettings
 import com.rsvpnano.models.PendingUpload
 import com.rsvpnano.models.RememberedNano
-import com.rsvpnano.models.needsArticleFetch
+import com.rsvpnano.library.needsArticleFetch
 import com.rsvpnano.persistence.AppSettingsStore
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CancellationException
