+++ companion/shared/src/commonMain/kotlin/com/rsvpnano/ui/RsvpNanoSharedApp.kt
@@ -68,11 +68,18 @@ import androidx.navigationevent.NavigationEventInfo
 import androidx.navigationevent.compose.NavigationBackHandler
 import androidx.navigationevent.compose.rememberNavigationEventState
 import com.rsvpnano.app.CompanionNotice
-import com.rsvpnano.app.NanoConnectionTransport
-import com.rsvpnano.app.NanoEndpoint
+import com.rsvpnano.connection.NanoConnectionTransport
+import com.rsvpnano.connection.NanoEndpoint
 import com.rsvpnano.converters.RsvpSupportedFileTypes
 import com.rsvpnano.models.PendingUpload
-import com.rsvpnano.models.needsArticleFetch
+import com.rsvpnano.library.needsArticleFetch
+import com.rsvpnano.presentation.CompanionPresenter
+import com.rsvpnano.presentation.CompanionUiState
+import com.rsvpnano.presentation.BookJob
+import com.rsvpnano.ui.library.*
+import com.rsvpnano.ui.settings.SETTINGS_INDEX_HELP
+import com.rsvpnano.ui.settings.SettingsDestination
+import com.rsvpnano.ui.settings.SettingsScreen
 import io.github.vinceglb.filekit.name
 import io.github.vinceglb.filekit.readBytes
 import io.github.vinceglb.filekit.dialogs.FileKitType
